Time: 8 ms (97.74%), Space: 15.4 MB (86.27%) - LeetHub

+++ 1517-restore-the-array/1517-restore-the-array.cpp
@@ -4,7 +4,7 @@ class Solution {
 public:
     int numberOfArrays(string s, int k) {
         int n = s.size();
-        vector<long long> dp(n+1, 0);
+        vector<int> dp(n+1, 0);
         dp[n] = 1;
 
         for (int i = n-1; i >= 0; i--) {
